refactor: tweak the One True Crate until it compiles and runs
