+++ src/lib.rs
@@ -1,36 +1,29 @@
-pub mod atmosphere;
-pub mod constants;
-pub mod core;
-pub mod forces;
-pub mod format;
-pub mod geometry;
-pub mod grid;
-pub mod ideal_gas;
-pub mod time;
-pub mod fluid_volume;
+pub mod physics;
+pub mod ui;
+pub mod scenario;
 
-pub use uom as units;
+use bevy::prelude::*;
 
-pub mod prelude {
-    pub use crate::{
-        atmosphere::Atmosphere,
-        core::{BuoyPlugin, SimState},
-        forces::{drag, scale_gravity},
-        grid::{Precision, RootGrid, GRID_CELL_EDGE_LENGTH_METERS},
-        ideal_gas::{GasSpecies, IdealGas},
-        fluid_volume::{FluidVolume, FluidVolumeCell, FluidVolumeGrid},
-    };
-    pub use uom::si::{
-        f32::{Mass, MassDensity, MolarMass, Pressure, ThermodynamicTemperature, Volume},
-        mass::kilogram,
-        mass_density::kilogram_per_cubic_meter,
-        molar_mass::kilogram_per_mole,
-        pressure::pascal,
-        thermodynamic_temperature::kelvin,
-        volume::cubic_meter,
-    };
+
+/// The plugin that handles the overall simulation state.
+pub struct BuoyPlugin;
+
+impl Plugin for BuoyPlugin {
+    fn build(&self, app: &mut App) {
+        app.init_state::<SimState>();
+        app.add_plugins((
+            #[cfg(feature = "render")]
+            ui::plugin,
+            physics::plugin,
+            scenario::plugin,
+        ));
+    }
 }
 
-pub use core::BuoyPlugin;
-
-pub use fluid_volume::{FluidVolumeBuilder, FluidVolumeCell, DefaultFluidVolumeSettings};
+#[derive(States, Debug, Default, Clone, Copy, Hash, PartialEq, Eq)]
+pub enum SimState {
+    Stopped,
+    #[default]
+    Running,
+    Faulted,
+}
