+++ src/main.rs
@@ -1,0 +1,10 @@
+use bevy::prelude::*;
+use buoy::BuoyPlugin;
+
+fn main() {
+    App::new()
+        .add_plugins((DefaultPlugins,
+            BuoyPlugin,
+        ))
+        .run();
+}
