+++ src/physics/geometry.rs
@@ -1,6 +1,6 @@
 #![allow(dead_code)]
 
-use crate::constants::PI;
+use super::constants::PI;
 
 pub fn sphere_volume(radius: f32) -> f32 {
     (4.0 / 3.0) * PI * f32::powf(radius, 3.0)
