+++ src/scenario/mod.rs
@@ -1,0 +1,5 @@
+use bevy::prelude::*;
+
+pub(super) fn plugin(_app: &mut App) {
+    // nothing yet
+}
